Refactor user routes into their own controller
